diplomacy & prci: adjust crossingtype locations and type hierarchy
