+++ src/main/scala/diplomacy/Parameters.scala
@@ -4,7 +4,7 @@ package freechips.rocketchip.diplomacy
 
 import Chisel._
 import chisel3.util.ReadyValidIO
-import freechips.rocketchip.util.{ShiftQueue, RationalDirection, FastToSlow, AsyncQueueParams, CreditedDelay}
+import freechips.rocketchip.util.{ShiftQueue}
 
 /** Options for describing the attributes of memory regions */
 object RegionType {
@@ -298,29 +298,6 @@ object TriStateValue
 {
   implicit def apply(value: Boolean): TriStateValue = TriStateValue(value, true)
   def unset = TriStateValue(false, false)
-}
-
-/** Enumerates the types of clock crossings generally supported by Diplomatic bus protocols  */
-sealed trait ClockCrossingType
-{
-  def sameClock = this match {
-    case _: SynchronousCrossing | _: CreditedCrossing => true
-    case _ => false
-  }
-}
-
-case object NoCrossing // converts to SynchronousCrossing(BufferParams.none) via implicit def in package
-case class SynchronousCrossing(params: BufferParams = BufferParams.default) extends ClockCrossingType
-case class RationalCrossing(direction: RationalDirection = FastToSlow) extends ClockCrossingType
-case class AsynchronousCrossing(depth: Int = 8, sourceSync: Int = 3, sinkSync: Int = 3, safe: Boolean = true, narrow: Boolean = false) extends ClockCrossingType
-{
-  def asSinkParams = AsyncQueueParams(depth, sinkSync, safe, narrow)
-}
-case class CreditedCrossing(sourceDelay: CreditedDelay, sinkDelay: CreditedDelay) extends ClockCrossingType
-
-object CreditedCrossing {
-  def apply(delay: CreditedDelay): CreditedCrossing = CreditedCrossing(delay, delay.flip)
-  def apply(): CreditedCrossing = CreditedCrossing(CreditedDelay(1, 1))
 }
 
 trait DirectedBuffers[T] {
