+++ src/main/scala/prci/ClockDomain.scala
@@ -4,10 +4,7 @@ import chisel3._
 import freechips.rocketchip.config.Parameters
 import freechips.rocketchip.diplomacy._
 
-abstract class ClockDomain(implicit p: Parameters) extends LazyModule
-  with LazyScope
-  with HasClockDomainCrossing
-{
+abstract class Domain(implicit p: Parameters) extends LazyModule with HasDomainCrossing {
   def clockBundle: ClockBundle
 
   lazy val module = new LazyRawModuleImp(this) {
@@ -23,6 +20,8 @@ abstract class ClockDomain(implicit p: Parameters) extends LazyModule
   }
 }
 
+abstract class ClockDomain(implicit p: Parameters) extends Domain with HasClockDomainCrossing
+
 class ClockSinkDomain(val clockSinkParams: ClockSinkParameters)(implicit p: Parameters) extends ClockDomain
 {
   def this(take: Option[ClockParameters] = None, name: Option[String] = None)(implicit p: Parameters) = this(ClockSinkParameters(take = take, name = name))
@@ -36,3 +35,5 @@ class ClockSourceDomain(val clockSourceParams: ClockSourceParameters)(implicit p
   val clockNode = ClockSourceNode(Seq(clockSourceParams))
   def clockBundle = clockNode.out.head._1
 }
+
+abstract class ResetDomain(implicit p: Parameters) extends Domain with HasResetDomainCrossing
